comment-rate-service/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
requests==2.32.5
